pyproject.toml
[tool.poetry]
name = "budget-manager"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
pillow = "10.1.0"
psycopg2 = "2.9.9"
asgiref = "3.7.2"
autopep8 = "2.0.4"
blinker = "1.7.0"
certifi = "2023.7.22"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
dash-core-components = "2.0.0"
dash-html-components = "2.0.0"
dash-table = "5.0.0"
django-appconf = "1.0.5"
django-compressor = "4.4"
django-debug-toolbar = "4.2.0"
django-environ = "0.11.2"
django-filter = "23.3"
django-silk = "5.0.4"
djangorestframework = "3.14.0"
everapi = "0.1.1"
factory-boy = "3.3.0"
faker = "19.13.0"
flask = "3.0.0"
freecurrencyapi = "0.1.0"
gprof2dot = "2022.7.29"
idna = "3.4"
importlib-metadata = "6.8.0"
inflection = "0.5.1"
iniconfig = "2.0.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
markupsafe = "2.1.3"
memory-profiler = "0.61.0"
nest-asyncio = "1.5.8"
packaging = "23.2"
pandas = "2.1.2"
plotly = "5.18.0"
pluggy = "1.3.0"
psutil = "5.9.6"
pycodestyle = "2.11.1"
pytest-django = "4.7.0"
python-dateutil = "2.8.2"
pytz = "2023.3.post1"
rcssmin = "1.1.1"
requests = "2.31.0"
retrying = "1.3.4"
rjsmin = "1.2.1"
setuptools = "68.2.2"
six = "1.16.0"
sqlparse = "0.4.4"
tenacity = "8.2.3"
typing-extensions = "4.8.0"
tzdata = "2023.3"
urllib3 = "2.0.7"
werkzeug = "3.0.1"
zipp = "3.17.0"
numpy = "^1.26.3"
pytesseract = "^0.3.10"
xhtml2pdf = "^0.2.14"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
